Created composite of reports, moved classes to its packages.
