--- exc1/src/main/java/dgodek/company/Tester.java
@@ -1,10 +1,0 @@
-package dgodek.company;
-
-/**
- * Created by matematyk60 on 23.07.17.
- */
-public class Tester extends Developer {
-    public Tester(String name) {
-        super(name, Role.TESTER);
-    }
-}
