+++ exc1/src/main/java/dgodek/company/employee/Manager.java
@@ -1,4 +1,4 @@
-package dgodek.company;
+package dgodek.company.employee;
 
 /**
  * Created by matematyk60 on 23.07.17.
