+++ exc1/src/main/java/dgodek/company/exceptions/NoSuchHiredEmployeeException.java
@@ -1,4 +1,4 @@
-package dgodek.company;
+package dgodek.company.exceptions;
 
 public class NoSuchHiredEmployeeException extends RuntimeException {
