+++ exc1/src/main/java/dgodek/company/exceptions/TooManyEmployeesException.java
@@ -1,4 +1,4 @@
-package dgodek.company;
+package dgodek.company.exceptions;
 
 /**
  * Created by matematyk60 on 23.07.17.
