+++ exc1/src/main/java/dgodek/company/report/Report.java
@@ -1,0 +1,5 @@
+package dgodek.company.report;
+
+public interface Report {
+    public void print();
+}
